HSEARCH-1188 Dependency commons-lang is required by the Elasticsearch tests: make it an explicit dependency

+++ orm/src/test/java/org/hibernate/search/test/query/dsl/embedded/DslEmbeddedSearchTest.java
@@ -83,9 +83,11 @@ public class DslEmbeddedSearchTest extends SearchTestBase {
 	@Override
 	@After
 	public void tearDown() throws Exception {
-		s.clear();
-		deleteAll( s, ContainerEntity.class );
-		s.close();
+		if ( s != null ) {
+			s.clear();
+			deleteAll( s, ContainerEntity.class );
+			s.close();
+		}
 		super.tearDown();
 	}
